Uso de Upload de fotos e JWT

+++ routes/api.php
@@ -4,6 +4,10 @@ use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
 use App\Http\Controllers\ApiController;
 use App\Http\Controllers\AuthController;
+use Illuminate\Support\Facades\Validator;
+use Illuminate\Support\Facades\Storage;
+use Illuminate\Support\Facades\Auth;
+
 //use App\Http\Controllers\{ LoginController, UserController, ContatoController,
    // PostController, ConteudoController
 //};
@@ -66,9 +70,7 @@ Route::get('/ping', function(){
        'pong' => true];
 
 });
-Route::get('/unautheticated', function(){
-return ['error' => 'Usuário não logado!'];
-})->name('login');
+
 // CRUD do TODO
 
 
@@ -83,14 +85,55 @@ return ['error' => 'Usuário não logado!'];
 // Put /todo/2 = Atualizar uma tarefa no sistema
 // Delete / todo/2 = Deletar uma tarefa no sistema
 
-
+Route::get('unauthenticated',  function(){
+    return ['error' => 'Usuário não está logado']; 
+})->name('login');
 
 Route::post('/user',[AuthController::class, 'create']);
-Route::post('/auth', [AuthController::class, 'login']);
+Route::post('/auth/login', [AuthController::class, 'login']);
+Route::middleware('auth:api')->post('/auth/logout', [AuthController::class, 'logout']);
+
+Route::middleware('auth:api')->get('/auth/me', [AuthController::class, 'me']);
 
 
-Route::middleware('auth:sanctum')->post('/todo', [ApiController::class, 'createTodo']);
+Route::middleware('auth:api')->post('/todo', [ApiController::class, 'createTodo']);
 Route::get('/todos', [ApiController::class, 'readAllTodos']);
 Route::get('/todo/{id}', [ApiController::class, 'readTodo']);
-Route::put('/todo/{id}', [ApiController::class, 'updateTodo']);
-Route::delete('/todo/{id}', [ApiController::class, 'deleteTodo']);
+Route::middleware('auth:api')->put('/todo/{id}', [ApiController::class, 'updateTodo']);
+Route::middleware('auth:api')->delete('/todo/{id}', [ApiController::class, 'deleteTodo']);
+
+Route::post('/upload', function(Request $request){
+    $array = ['error' => ''];
+
+    $rules = [
+        'name' => 'required|min:2',
+        'foto' => 'required|mimes:jpg,png'
+    ];
+
+     $validator = Validator::make($request->all(), $rules );
+
+        if($validator->fails()){
+            $array['error'] = $validator->messages();
+
+            return $array;
+        }
+     
+
+    if($request->hasFile('foto')){
+
+      if($request->file('foto')->isValid()) {
+        
+        $foto =  $request->file('foto')->store('public');
+
+        $url = asset(Storage::url($foto));
+
+        $array['url'] = $url;
+
+      }
+    
+    } else {
+        $array['error'] = 'Não foi enviado arquivo.';
+    }
+
+    return $array;
+});
